requirements.txt
pandas
kafka-python
minio
scikit-learn
joblib
streamlit
requests